// src/types.ts
export enum ELocale {
    ENGLISH = 'en-US',
    VIETNAM = 'vi-VN',
    JAPAN = 'ja-JP',
    THAILAN = 'th-TH',
    INDONESIA = 'id-ID',
    MALAYSIA = 'ms-MY',
    CHINA = 'zh-CN',
    KHMER = 'km',
}

export enum EDeviceType {
    DESKTOP = 'Desktop',
    MOBILE = 'Mobile',
}

export type TCoin = 'USDT' | 'IDA' | 'IDAA' |'IMD'
export type TCoinInvest = 'USDT' | 'IMD'